+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 * Added build_di_container() method to class Main
 * Added class-components.php helper class for easier component rendering
 * Added ability to wrap components with parent class on render
+* class-main.php - Added build_di_container() method.
+* class-blocks.php - Added custom filter `block-attributes-override` to be able to override attributes depending on the post type.
+* class-invalid-block.php - Fixed error msg.
 
 ## [2.0.7] - 2020-01-29